ease back on resolution of tests
The tested values from Dec 2020 differ from those of Jan 2021 in the 3rd
or 4th place after the decimal place in the seconds field.  This is a
very small difference in practical terms, and I suppose it's not
surprising that there might be shifts in the formulae over time, as new
data become available.